+++ R/utilities.R
@@ -61,13 +61,13 @@ flattenList <- function(x, type) {
 
   if (is.list(x)) {
     x <- switch(type,
-      "character" = purrr::flatten_chr(x),
+      "character" = purrr::list_c(x, ptype = "c"),
       "numeric" = ,
       "real" = ,
-      "double" = purrr::flatten_dbl(x),
-      "integer" = purrr::flatten_int(x),
-      "logical" = purrr::flatten_lgl(x),
-      purrr::flatten(x)
+      "double" = purrr::list_c(x, ptype = 1.0),
+      "integer" = purrr::list_c(x, ptype = 1),
+      "logical" = purrr::list_c(x, ptype = TRUE),
+      purrr::list_flatten(x)
     )
   }
 
@@ -110,4 +110,63 @@ toMissingOfType <- function(x, type) {
   }
 
   return(x)
+}
+
+#' Computes logarithm of a number or of a vector of numbers and handles zeros while
+#' substituting all values below `epsilon` by `epsilon`.
+#'
+#' @param x A numeric or a vector of numerics.
+#' @param base a positive or complex number: the base with respect to which logarithms are computed. Defaults to e = exp(1).
+#' @param epsilon A very small number which is considered as threshold below which
+#' all values are treated as `epsilon`. Allows computation of `log` close to 0.
+#' Default value is `getOSPSuiteUtilsSetting("LOG_SAFE_EPSILON")`.
+#'
+#' @return `log(x, base = base)` for `x > epsilon`, or `log(epsilon, base = base)`,
+#' or `NA_real_` for `NA` elements.
+#' @export
+#'
+#' @examples
+#' inputVector <- c(NA, 1, 5, 0, -1)
+#' logSafe(inputVector)
+logSafe <- function(x, base = exp(1), epsilon = ospsuiteUtilsEnv$LOG_SAFE_EPSILON) {
+  x <- sapply(X = x, function(element) {
+    element <- ospsuite.utils::toMissingOfType(element, type = "double")
+    if (is.na(element)) {
+      return(NA_real_)
+    } else if (element < epsilon) {
+      return(log(epsilon, base = base))
+    } else {
+      return(log(element, base = base))
+    }
+  })
+
+  return(x)
+}
+
+#' Safe fold calculation
+#'
+#' @description
+#' Calculates `x / y` while substituting values below `epsilon` (for x and y) by `epsilon`.
+#' `x` and `y` must be of the same length
+#'
+#'
+#' @param x A numeric or a vector of numerics.
+#' @param y A numeric or a vector of numerics.
+#' @param epsilon A very small number which is considered as threshold below which
+#' all values are treated as `epsilon`. Allows computation of fold changes for values close to 0.
+#' Default value is `getOSPSuiteUtilsSetting("LOG_SAFE_EPSILON")`.
+#'
+#' @return A vector with `x / y`.
+#' @export
+#'
+#' @examples
+#' inputX <- c(NA, 1, 5, 0, -1)
+#' inputY <- c(1, -1, NA, 0, -1)
+#' folds <- foldSafe(inputX, inputY)
+foldSafe <- function(x, y, epsilon = ospsuiteUtilsEnv$LOG_SAFE_EPSILON) {
+  validateIsSameLength(x, y)
+  x[x <= epsilon] <- epsilon
+  y[y <= epsilon] <- epsilon
+
+  return(x / y)
 }